README File
README.txt
The a038_rc_wifi_991.aia and its associated extension UrsAI2UDPv3.aix
were incorporated in the a038_gamepad_otg_wifi.aia.  The a038_rc_wifi_991
provides the required WIFI functions and properties.  

This folder provides the aia and aix for the WIFI as design maturity artifacts
that may be useful in future projects.

They are present "as-is" with no plans to support future changes.