Revised downard the granularity of control
Road test showed that the number of segments (24 for 360) was ideal but prohibitive in performance and caused excessive delay.  Instead a "sloppy" differential style steering supported the control of the vehicle within the expectations of a typical user who would be approaching a target of interest who might also be moving.  This sloppy steering resulted in the evolution of six "workable" segments.  See http://abbottanp.com/artifacts/squirrelChaser/page2.html#page2 Helpful Images showing Implemented as Shown Six Segment image and Mapped to Switch-Case image.